スポイトを追加してリファクタ
ボタンやラベルなどの生成をジェネレータとして別のファイルに記述してコンパイル時間を短縮する
