+++ LyapunovViewer/GeneratorContext.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <Windows.h>
+#include "Resource.h"
+
+void GenerateSpoitButton(HWND hWnd, LPARAM& lParam);
+void GenerateLabels(HDC& hdc);
