+++ LyapunovViewer/Resource.h
@@ -16,6 +16,10 @@
 #define BUTTON_ID_OPEN_FILE_DIALOG 200
 #define BUTTON_ID_SAVE_FILE_DIALOG 201
 #define BUTTON_ID_RUN_LYANUNOV 202
+#define BUTTON_ID_SPOIT_AMAX 203
+#define BUTTON_ID_SPOIT_AMIN 204
+#define BUTTON_ID_SPOIT_BMAX 205
+#define BUTTON_ID_SPOIT_BMIN 206
 #ifndef IDC_STATIC
 #define IDC_STATIC				-1
 #endif
