--- LyapunovViewer/Utility.h
@@ -1,9 +1,0 @@
-#pragma once
-#include <windows.h>
-
-void LabelOut(HDC& hdc, const int x, const int y, LPCWSTR label)
-{
-    TextOut(hdc, x, y, label, lstrlen(label));
-    MoveToEx(hdc, x, y + 24, NULL);
-    LineTo(hdc, 300, y + 24);
-}
